Make GetProgress resturn int in range 0-100.
float progress caused problems with equality to 1.

+++ filerw.h
@@ -43,7 +43,7 @@ public:
 	void TestLoop();
 	void InitScene();
 	void UpdateScene();
-	qreal GetProgress();
+	int GetProgress();
 
 	FileRWResults results_write;	// results of write test
 	FileRWResults results_read;		// results of read test
