+++ readrnd.h
@@ -50,7 +50,7 @@ public:
 	QList<ReadRndResult> reference;
 
 	// test progress
-	qreal GetProgress();
+	int GetProgress();
 
 	QDomElement WriteResults(QDomDocument &doc);	// writes results of test to XML
 	void RestoreResults(QDomElement &root, bool reference = false);			// reads results from XML document
